requirements.txt
djangorestframework
django-environ
dj-rest-auth
dj-rest-auth[with_social]
djangorestframework-simplejwt
